requirements.txt
numpy
scipy
matplotlib
scikit-image
PyYAML
torch
wandb
pybullet

-e ./smp_manifold_learning-master
